Added some basic json state file writing.
Much more is needed. Like, to read from this file every so often.
I also create a Scum object to track the data around a given scummy ip.
This made it possilbe to provide a list of files that we are not going
to block because of the DoNotBlock bool value for each Scum.

NOTE: Do we even care about managing this in a json file now? How much
memory are we talking about? I bet it is in the Kb of memory. I bet it
would fit nicely without putting it into a file.